feat: circom module and cli options
