+++ .gitignore
@@ -18,4 +18,4 @@ Cargo.lock
 .env
 
 # Output
-*.json
+output/
