+++ src/circom/execution.rs
@@ -1,0 +1,44 @@
+use circom_compiler::compiler_interface::VCP;
+use circom_constraint_generation::{build_circuit, BuildConfig};
+use circom_program_structure::program_archive::ProgramArchive;
+
+pub struct ExecutionConfig {
+    pub r1cs: String,
+    pub sym: String,
+    pub json_constraints: String,
+    pub json_substitutions: String,
+    pub no_rounds: usize,
+    pub flag_s: bool,
+    pub flag_f: bool,
+    pub flag_p: bool,
+    pub flag_old_heuristics: bool,
+    pub flag_verbose: bool,
+    pub inspect_constraints_flag: bool,
+    pub sym_flag: bool,
+    pub r1cs_flag: bool,
+    pub json_substitution_flag: bool,
+    pub json_constraint_flag: bool,
+    pub prime: String,
+}
+
+pub fn execute_project(
+    program_archive: ProgramArchive,
+    config: ExecutionConfig,
+) -> Result<VCP, ()> {
+    let build_config = BuildConfig {
+        no_rounds: config.no_rounds,
+        flag_json_sub: config.json_substitution_flag,
+        json_substitutions: config.json_substitutions,
+        flag_s: config.flag_s,
+        flag_f: config.flag_f,
+        flag_p: config.flag_p,
+        flag_verbose: config.flag_verbose,
+        inspect_constraints: config.inspect_constraints_flag,
+        flag_old_heuristics: config.flag_old_heuristics,
+        prime: config.prime,
+    };
+
+    let (_, vcp) = build_circuit(program_archive, build_config)?;
+
+    Result::Ok(vcp)
+}
