+++ src/lib.rs
@@ -2,8 +2,8 @@
 //!
 //! This library provides the functionality to convert a Circom program into an arithmetic circuit.
 
+pub mod circom;
 pub mod circuit;
-pub mod compiler;
 pub mod process;
 pub mod program;
 pub mod runtime;
